// src/app/interfaces/ITaskDetail.ts
export interface ITaskDetail {
  displayDescription?: string;
  executedBy?: string;
  rootContainerId?: string;
  assigned_date?: string;
  displayName?: string;
  executedBySubstitute?: string;
  dueDate?: string;
  description?: string;
  type?: string;
  priority?: string;
  actorId?: string;
  processId?: string;
  caseId?: string;
  name?: string;
  reached_state_date?: string;
  rootCaseId?: string;
  id?: string;
  state?: string;
  parentCaseId?: string;
  last_update_date?: string;
  assigned_id?: string;
}